requirements.txt
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
psycopg2==2.9.5
python-dotenv==0.21.1
